Fold phoenix adapter into importer package
