+++ plugins/packages/phoenix-importer/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add the Arize Phoenix importer-backed adapter, replacing the `kitaru-phoenix` distribution.
+
 ## 0.1.0
 
 - Add the Arize Phoenix JSON and JSONL trace importer.
