+++ plugins/packages/phoenix-importer/pyproject.toml
@@ -22,7 +22,12 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 keywords = ["ai-agents", "arize-phoenix", "kitaru", "observability", "traces"]
-dependencies = ["kitaru>=0.22.0"]
+dependencies = [
+    "arize-phoenix-client>=2.0,<4",
+    "kitaru>=0.23.0",
+    "opentelemetry-api>=1.12",
+    "opentelemetry-sdk>=1.12",
+]
 
 [project.urls]
 Homepage = "https://kitaru.ai"
