--- plugins/packages/phoenix/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Changelog
-
-## 0.1.0
-
-- Initial release of the Arize Phoenix importer-backed adapter.
