+++ plugins/pyproject.toml
@@ -12,7 +12,6 @@ kitaru-langgraph = { workspace = true }
 kitaru-langsmith-importer = { workspace = true }
 kitaru-logfire-importer = { workspace = true }
 kitaru-openai-agents = { workspace = true }
-kitaru-phoenix = { workspace = true }
 kitaru-phoenix-importer = { workspace = true }
 kitaru-pydantic-ai = { workspace = true }
 
@@ -25,7 +24,7 @@ dev = [
     "kitaru-langsmith-importer",
     "kitaru-logfire-importer",
     "kitaru-openai-agents",
-    "kitaru-phoenix",
+    "kitaru-phoenix-importer",
     "kitaru-pydantic-ai[openai]",
     "pytest>=8.0",
     "pytest-asyncio>=1.4.0",
@@ -56,7 +55,6 @@ known-first-party = [
     "kitaru_logfire_importer",
     "kitaru_langsmith_importer",
     "kitaru_openai_agents",
-    "kitaru_phoenix",
     "kitaru_phoenix_importer",
     "kitaru_pydantic_ai",
 ]
